fix rafraichissement widgets + lancement du service de lecture
